Fixed folders - read desc
So, I messed up my workspace because I put my workspace folder in a seperate folder to the root folder. So fred's changes were directly to the root repository and mine were to the "Powerboard25" folder inside the root folder. I removed my folder and my my workspace the in the root folder. the IOC and all other files are from fred's (which should be very similar to mine, just a little older) I just transfered over the ltc2990 files